requirements.txt
streamlit
opencv-python-headless
mediapipe
google-generativeai
Pillow
fpdf2
python-bidi